Fix warnings on LP64 platforms
    The code assumed "sizeof(int) == sizeof(void*)".

+++ mpich2/test/mpi/threads/pt2pt/multisend2.c
@@ -25,7 +25,7 @@ static int nthreads = -1;
 MTEST_THREAD_RETURN_TYPE run_test_send(void *arg)
 {
     int    cnt, j, *buf;
-    int    thread_num = (int)arg;
+    int    thread_num = (int)(size_t)arg;
     double t;
 
     for (cnt=1; cnt < MAX_CNT; cnt = 2*cnt) {
@@ -88,7 +88,7 @@ int main(int argc, char ** argv)
     if (rank == 0) {
 	nthreads = nprocs-1;
 	for (i=1; i<nprocs; i++) 
-	    MTest_Start_thread( run_test_send,  (void *)i );
+	    MTest_Start_thread( run_test_send,  (void *)(size_t)i );
 	MTest_Join_threads( );
     }
     else {
